+++ tests/service.py
@@ -34,6 +34,10 @@ def create_payment_method_id():
     return f'las:payment-method:{uuid4().hex}'
 
 
+def create_prediction_id():
+    return f'las:prediction:{uuid4().hex}'
+
+
 def create_deployment_environment_id():
     return f'las:deployment-environment:{uuid4().hex}'